refactor(reducers): move `details-pane-expanded`
