+++ app/renderer/src/reducers/details-pane-expanded/index.js
@@ -1,0 +1,1 @@
+export { default } from './details-pane-expanded';
